requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
boto==2.49.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cloudmersive-ocr-api-client==3.0.1
discord.py==1.5.1
google-api-core==1.23.0
google-auth==1.23.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
libcst==0.3.15
multidict==4.7.6
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
yarl==1.5.1